bug fixes in DataEvaluation
