+++ src/tp3/src/tp3/Rosbag_Analysis.py
@@ -1,15 +1,17 @@
 import rosbag
 import genpy.rostime
 import numpy as np
+import sys
+sys.path.append('../src/tp3')
+import DataEvaluation_module as de
 
 # These functions can be called by the file "Rosbag_Analysis_main.py" #
 
 TOPIC = 'objectlist'
-IoU_threshold = 0.0     # tbd
 
 class Rosbag_Analysis:
   
-    ### method for getting count of objects contained in rosbag ###
+    ### method for getting available object IDs contained in rosbag ###
     @staticmethod
     def getObjectIDs(bagfile):
         bag = rosbag.Bag(bagfile)
@@ -20,9 +22,24 @@ class Rosbag_Analysis:
                 if i.obj_id not in array_ids:
                     array_ids.append(i.obj_id)
                
-        np.sort(array_ids)
+        return_array = np.sort(array_ids)
         bag.close()
-        return array_ids
+        return return_array
+        
+    ### method for getting count of objects contained in rosbag summed up over all messages ###
+    @staticmethod
+    def getObjectCountTotal(bagfile):
+        bag = rosbag.Bag(bagfile)
+        objects_msg = []
+        count_objects_total = 0
+       
+        for topic, msg, t in bag.read_messages(topics=[TOPIC]):
+            for i in msg.obj_list:
+                objects_msg.append(i)
+                count_objects_total += len(objects_msg)
+               
+        bag.close()
+        return count_objects_total
     
     
     ### method for getting array of raw values of all frames in a rosbag ###
@@ -49,7 +66,10 @@ class Rosbag_Analysis:
             # to find array index of target object ID
             for i in msg.obj_list:
                 if i.obj_id == obj_id_target:
-                    array_values.append(getattr(getattr(msg.obj_list[index_counter], category), attribute))
+                    if category == "": # branch with one level in object list tree (e.g. 'obj_id')
+                        array_values.append(getattr(msg.obj_list[index_counter], attribute))
+                    else:
+                        array_values.append(getattr(getattr(msg.obj_list[index_counter], category), attribute))
                     array_timestamps.append((float)((t.__sub__(startTime)).__str__()) / 1000000)  #timestamp in milli seconds
                     break
                 else: index_counter += 1
@@ -65,22 +85,22 @@ class Rosbag_Analysis:
     
     
     @staticmethod
-    def getAdvancedData(bagfile1, bagfile2, obj_id_target_gt, category, attribute, operation):
+    def getAdvancedData(bagfile1, bagfile2, obj_id_target_gt, category, attribute, operation, IoU_threshold):
     
-        (timestamps_gt, values_gt, values_cam) = Rosbag_Analysis.getCalculationValues(bagfile1, bagfile2, obj_id_target_gt, category, attribute)
+        (timestamps_gt, values_gt, values_cam) = Rosbag_Analysis.getCalculationValues(bagfile1, bagfile2, obj_id_target_gt, category, attribute, IoU_threshold)
         
         ### operation ###
         if operation == "difference":
             array_result = np.subtract(values_gt, values_cam)
-        
+            
         return (timestamps_gt, array_result)
      ######################  
-
-
+     
+     
      
     @staticmethod
     #returns (timestamp_gt, values_gt, values_cam) 
-    def getCalculationValues(bagfile1, bagfile2, obj_id_target_gt, category, attribute):
+    def getCalculationValues(bagfile1, bagfile2, obj_id_target_gt, category, attribute, IoU_threshold):
         
         bag_gt = rosbag.Bag(bagfile1)
         bag_cam = rosbag.Bag(bagfile2)
@@ -90,15 +110,12 @@ class Rosbag_Analysis:
         
         startTimeRaw_cam= bag_cam.get_start_time()
         startTime_cam = genpy.rostime.Time.from_sec(startTimeRaw_cam)
-
         
-        array_msg_cam = []
-        
-        array_values_gt = []
         array_timestamps_gt = []
+        array_values_gt = []
         array_values_cam = []
-
-        # loop through GT messages
+        
+        # loop through GT messages/frames
         for topic, msg, t in bag_gt.read_messages(topics=[TOPIC]):
             bag_gt.read_messages()
             timestamp_gt = (float)((t.__sub__(startTime_gt)).__str__()) / 1000000   #normed timestamp GT in ms
@@ -108,21 +125,30 @@ class Rosbag_Analysis:
             index_counter = 0
             obj_id_index = 99
             
+            array_objects_gt = []
+            
             obj_cam = 0
             value_cam_found = 0
+            last_value_cam = 0
             
             # ID mapping:
             # loop through GT objects in message --> get right GT object from ID 
             for i in msg.obj_list:
+                #collect all GT objects
+                array_objects_gt.append(i)
+                
+                #get calculation values of target GT object with timestamp
                 if i.obj_id == obj_id_target_gt:      
                                 # timestamp       # GT object    
                     obj_gt = [timestamp_gt, msg.obj_list[index_counter]]
-                    array_values_gt.append(getattr(getattr(obj_gt[1], category), attribute))
-                    break
+                    if category == "": # branch with one level in object list tree (e.g. 'obj_id')
+                        array_values_gt.append(getattr(obj_gt[1], attribute))
+                    else:
+                        array_values_gt.append(getattr(getattr(obj_gt[1], category), attribute))
                 else: index_counter += 1 # next object
-           
+      
             # finding concerning CAM object 
-            # loop through CAM messages
+            # loop through CAM messages/frames to find message concerning to GT frame (in time) 
             for topic, msg, t in bag_cam.read_messages(topics=[TOPIC]):
                 bag_cam.read_messages()
                 timestamp_cam = (float)((t.__sub__(startTime_cam)).__str__()) / 1000000   #normed timestamp CAM in ms
@@ -133,28 +159,32 @@ class Rosbag_Analysis:
                 and (timestamp_cam < (timestamp_gt + 500))):
                     
                     # position mapping: 
-                    # loop through CAM objects in message --> compare all CAM objects with GT object
-                    if value_cam_found == 0:
+                    # loop through CAM objects in message --> proceed IoU evaluation
+                    if value_cam_found == 0: # for time mapping # PROBLEMATISCH???
                         for i in msg.obj_list:
-                            prop_IoU = 0
                                             #index 0: timest.  #index 1: object
                             obj_cam = [timestamp_cam, i]
                           
                             ###
-                            #   IoU testing: each CAM object compared to current GT object
-                            #prop_IoU = testIoU(obj_gt[1], obj_cam[1])
+                            #IoU testing: each CAM object compared to list of GT objects
+                            (evaluation, index_TP) = de.det_TP_FP_mm(array_objects_gt, obj_cam[1], IoU_threshold)
                             ###
-                            
-                            if prop_IoU >= IoU_threshold:
-                                last_value_cam = getattr(getattr(obj_cam[1], category), attribute)
-                                array_values_cam.append(getattr(getattr(obj_cam[1], category), attribute))
+
+                            # TP (true positive) case - you got a match:
+                            if evaluation == 0 and array_objects_gt[index_TP].obj_id == obj_id_target_gt:
+                                if category == "": # branch with one level in object list tree (e.g. 'obj_id')
+                                    last_value_cam = getattr(obj_cam[1], attribute)
+                                    array_values_cam.append(getattr(obj_cam[1], attribute))
+                                else:
+                                    last_value_cam = getattr(getattr(obj_cam[1], category), attribute)
+                                    array_values_cam.append(getattr(getattr(obj_cam[1], category), attribute))
                                 value_cam_found = 1
                                 break
-                        
+
             if value_cam_found == 0:
                 array_values_cam.append(last_value_cam)
-                print('insert last value: ')
-                print(last_val_cam)
+                #print('insert last value: ')
+                #print(last_val_cam)
         
         bag_gt.close()
         bag_cam.close()
@@ -166,9 +196,176 @@ class Rosbag_Analysis:
         return (array_timestamps_gt, array_values_gt, array_values_cam)
     ######################        
      
-     
+    @staticmethod
+    def getEvaluation(bagfile1, bagfile2, IoU_threshold):
+        
+        bag_gt = rosbag.Bag(bagfile1)
+        bag_cam = rosbag.Bag(bagfile2)
+		       
+        startTimeRaw_gt= bag_gt.get_start_time()
+        startTime_gt = genpy.rostime.Time.from_sec(startTimeRaw_gt)
+        
+        startTimeRaw_cam= bag_cam.get_start_time()
+        startTime_cam = genpy.rostime.Time.from_sec(startTimeRaw_cam)
+        
+        array_timestamps_gt = []
+        
+        # loop through GT messages/frames
+        for topic, msg, t in bag_gt.read_messages(topics=[TOPIC]):
+            bag_gt.read_messages()
+            timestamp_gt = (float)((t.__sub__(startTime_gt)).__str__()) / 1000000   #normed timestamp GT in ms
+            array_timestamps_gt.append(timestamp_gt)
+            
+            array_objects_gt = []
+            array_objects_cam = []
+            
+            count_TP = 0
+            count_FP = 0 
+            count_mm = 0
+            count_FN = 0
+           
+            array_TP = []
+            array_FP = []
+            array_FN = []
+            array_mm = []
+            array_precision = []
+            array_recall = []
+            
+            #collect all GT objects
+            for i in msg.obj_list:
+                array_objects_gt.append(i)
+
+            # finding concerning CAM object 
+            # loop through CAM messages/frames to find message concerning to GT frame (in time) 
+            for topic, msg, t in bag_cam.read_messages(topics=[TOPIC]):
+                bag_cam.read_messages()
+                timestamp_cam = (float)((t.__sub__(startTime_cam)).__str__()) / 1000000   #normed timestamp CAM in ms
+                
+                # time mapping: 
+                # cam message in range +/- 500 ms from gt message
+                if((timestamp_cam > (timestamp_gt - 500))
+                and (timestamp_cam < (timestamp_gt + 500))):
+                    
+                    # position mapping: 
+                    # loop through CAM objects in message --> proceed IoU evaluation
+                    for i in msg.obj_list:
+                        #collect all CAM objects
+                        array_objects_cam.append(i)
+                                        #index 0: timest.  #index 1: object
+                        #obj_cam = [timestamp_cam, i]
+                      
+                    ###
+                    #   IoU testing: each CAM object compared to list of GT objects
+                    array_evaluations = de.det_TP_FP_mm(array_objects_gt, array_objects_cam, IoU_threshold)
+                    ###
+                    
+                    # analysing evaluation results
+                    for row in array_evaluations:
+                    
+                        # TP (true positive) case - you got a match:
+                        if row[0] == 0:
+                            count_TP += 1
+                        
+                        # FP (false positive) case:
+                        elif row[0] == 1:
+                            count_FP += 1
+                        
+                        # mm (mismatch) case:
+                        elif row[0] == 2:
+                            count_mm += 1
+                  
+                    # test FN (false negative) case:                    
+                    (array_evaluation_FN) = de.isFN(array_objects_gt, array_objects_cam, IoU_threshold)
+                    
+                    for i in array_evaluation_FN:
+                        if i == True:
+                            count_FN += 1
+                    
+                # add values per GT frame
+                array_TP.append(count_TP)
+                array_FP.append(count_FP)
+                array_FN.append(count_FN) 
+                array_mm.append(count_mm)
+                array_precision.append(count_TP / (count_TP + count_FP))
+                array_recall.append(count_TP / (count_TP + count_FN))
+            
+        bag_gt.close()
+        bag_cam.close()
+        return (array_timestamps_gt, array_TP, array_FP, array_FN, array_mm, array_precision, array_recall)
+        
+    '''
+    @staticmethod
+    def getmm(bagfile1, bagfile2, IoU_threshold):
+        
+        (timestamp, TP, FP, FN, mm) = getEvaluation(bagfile1, bagfile2, IoU_threshold)
+        
+        return (timestamp, mm)
+    '''
     
-    ### example method - DEPRECATED ###
+    @staticmethod
+    def getTP(bagfile1, bagfile2, IoU_threshold):
+        
+        (timestamp, TP, FP, FN, mm, precision, recall) = Rosbag_Analysis.getEvaluation(bagfile1, bagfile2, IoU_threshold)
+        
+        return (timestamp, TP)
+    
+    @staticmethod
+    def getFP(bagfile1, bagfile2, IoU_threshold):
+        
+        (timestamp, TP, FP, FN, mm, precision, recall) = Rosbag_Analysis.getEvaluation(bagfile1, bagfile2, IoU_threshold)
+        
+        return (timestamp, FP)
+    
+    @staticmethod
+    def getFN(bagfile1, bagfile2, IoU_threshold):
+        
+        (timestamp, TP, FP, FN, mm, precision, recall) = Rosbag_Analysis.getEvaluation(bagfile1, bagfile2, IoU_threshold)
+        
+        return (timestamp, FN)
+    
+    @staticmethod
+    def getPrecision(bagfile1, bagfile2, IoU_threshold):
+        
+        (timestamp, TP, FP, FN, mm, precision, recall) = Rosbag_Analysis.getEvaluation(bagfile1, bagfile2, IoU_threshold)
+        
+        return (timestamp, precision)
+    
+    @staticmethod
+    def getRecall(bagfile1, bagfile2, IoU_threshold):
+    
+        (timestamp, TP, FP, FN, mm, precision, recall) = Rosbag_Analysis.getEvaluation(bagfile1, bagfile2, IoU_threshold)
+        
+        return (timestamp, recall)
+        
+    @staticmethod
+    def getFPPI(bagfile1, bagfile2, IoU_threshold):
+            
+        (timestamp, TP, FP, FN, mm, precision, recall) = Rosbag_Analysis.getEvaluation(bagfile1, bagfile2, IoU_threshold)
+        
+        FP_total = np.sum(FP)
+        FPPI = FP_total / len(timestamp)  #len(timestamp) == nFrames_gt
+        
+        return FPPI
+            
+    @staticmethod
+    def getMOTA(bagfile1, bagfile2, IoU_threshold):
+        
+        (timestamp, TP, FP, FN, mm, precision, recall) = Rosbag_Analysis.getEvaluation(bagfile1, bagfile2, IoU_threshold)
+            
+        FP_total = np.sum(FP)
+        FN_total = np.sum(FN)
+        mm_total = np.sum(mm)
+        count_objects_GT_total = Rosbag_Analysis.getObjectCountTotal(bagfile1)
+
+        MOTA = 1 - ((FN_total + FP_total + mm_total) / count_objects_GT_total)
+    
+    #@staticmethod
+    #def get MOTP
+    
+
+
+
+   ### example method - DEPRECATED ###
     ### get all x values of object 1 ( = index 1) ###
     ### following method is deprecated ###
